2-Back_End/Proyecto/Antawa_JWT/requirements.txt
django==3.2
djangorestframework
djangorestframework-simplejwt
